Rename user create and login mutation files
