+++ app/graphql/types/role_type.rb
@@ -1,0 +1,5 @@
+Types::RoleType = GraphQL::ObjectType.define do
+	name 'Role'
+
+	field :name, !types.String
+end
